LayoutTests/platform/ios-sim/fast/events/frame-tab-focus-expected.txt

This page tests tabbing between subframes. To test, click on this text to focus the main window. Then press Tab 7 times, then Shift-Tab 7 times, which should move focus forward and backward through all inputs and frames. Then press Option-Tab 11 times and Shift-Option-Tab 11 times, which should move focus forward and backward through all inputs, frames, and links.

        [tabindex of one] [tabindex of three] [tabindex of two] [tabindex of three]
Tabbing forward...

main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: window blurred
main window: window focused
main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred

Tabbing backward...

main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred
main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: window blurred
main window: window focused
main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred

Option-tabbing forward...

main window: A #0 (tabindex=1) focused
main window: A #0 (tabindex=1) blurred
main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: A #2 (tabindex=2) focused
main window: A #2 (tabindex=2) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred
main window: A #1 (tabindex=3) focused
main window: A #1 (tabindex=3) blurred
main window: A #3 (tabindex=3) focused
main window: A #3 (tabindex=3) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: window blurred
main window: window focused
main window: A #0 (tabindex=1) focused
main window: A #0 (tabindex=1) blurred
main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: A #2 (tabindex=2) focused
main window: A #2 (tabindex=2) blurred

Option-tabbing backward...

main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: A #3 (tabindex=3) focused
main window: A #3 (tabindex=3) blurred
main window: A #1 (tabindex=3) focused
main window: A #1 (tabindex=3) blurred
main window: INPUT #1 (tabindex=3) focused
main window: INPUT #1 (tabindex=3) blurred
main window: A #2 (tabindex=2) focused
main window: A #2 (tabindex=2) blurred
main window: INPUT #2 (tabindex=2) focused
main window: INPUT #2 (tabindex=2) blurred
main window: A #0 (tabindex=1) focused
main window: A #0 (tabindex=1) blurred
main window: window blurred
main window: window focused
main window: INPUT #3 (tabindex=0) focused
main window: INPUT #3 (tabindex=0) blurred
main window: INPUT #0 (tabindex=0) focused
main window: INPUT #0 (tabindex=0) blurred
main window: A #3 (tabindex=3) focused
main window: A #3 (tabindex=3) blurred

Test finished
